refactor(config): split key binding internals
